keeping track of the score

+++ lib/pages/topics/transformations/menus/beginner_level.dart
@@ -166,7 +166,7 @@ class BeginnerLs {
   String rating;
   String answer;
   String hint;
-  double timeLimit;
+  int timeLimit;
   List<Offset> qPoints;
   List<Offset> optionA;
   List<Offset> optionB;
